+++ frontend/src/app/app.routes.ts
@@ -2,14 +2,12 @@ import { Routes } from '@angular/router';
 
 export const routes: Routes = [
     { path: '', redirectTo: 'login', pathMatch: 'full' },
-    { path: '', loadChildren: () => import('./auth/auth-module').then(m => m.AuthModule)},
-    { path: 'dashboard', loadChildren: () => import('./dashboard/dashboard-module')
+    { path: '', loadChildren: () => import('./features/auth/auth-module').then(m => m.AuthModule)},
+    { path: 'dashboard', loadChildren: () => import('./features/dashboard/dashboard-module')
             .then(m => m.DashboardModule)
     },
     {path: 'unauthorized', loadChildren: () => import('./system-pages/system-pages-module')
         .then(m => m.SystemPagesModule)
     },
-    {path: 'user/dashboard', loadChildren: () => import('./user-dashboard/user-dashboard-module')
-        .then(m => m.UserDashboardModule)
-    }
+
 ];
