+++ frontend/src/app/features/auth/auth-routing-module.ts
@@ -1,12 +1,12 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
-import { Login as LoginComponent } from './login/login';
-import { isLoggedInGuard, isAdminAndManagerAuthed } from '../core/guards/auth-guard-guard';
+import { Login as LoginComponent } from './pages/login/login';
+import { isLoggedInGuard, isAdminAndManagerAuthed } from '../../core/guards/auth-guard-guard'
 
 const routes: Routes = [
   {path: 'login', component: LoginComponent}, 
-  {path: 'register', loadComponent: () => import('./register/register').then(c => c.Register)},
-  {path: 'registerManager', loadComponent: () => import('./register-manager/register-manager')
+  {path: 'register', loadComponent: () => import('./pages/register/register').then(c => c.Register)},
+  {path: 'registerManager', loadComponent: () => import('./pages/register-manager/register-manager')
       .then(m => m.RegisterManager), canActivate:[isLoggedInGuard, isAdminAndManagerAuthed]
   }
 ];
