+++ frontend/src/app/features/auth/pages/register-manager/register-manager.html
