+++ frontend/src/app/features/auth/pages/register-manager/register-manager.spec.ts
