+++ frontend/src/app/features/dashboard/awaiting-pickup/awaiting-pickup.ts
@@ -1,20 +1,24 @@
 import { Component, OnInit } from '@angular/core';
 import { Observable } from 'rxjs';
-import { UserDashboardService } from '../user-dashboard-service';
-import { NgIf, AsyncPipe } from '@angular/common';
+import { Dashboard } from '../dashboard';
+import { AsyncPipe, NgIf } from '@angular/common';
+import { AppConsole } from '../../../utils/app-console';
 
 @Component({
   selector: 'app-awaiting-pickup',
+  standalone: true,
   imports: [NgIf, AsyncPipe],
   templateUrl: './awaiting-pickup.html',
   styleUrl: './awaiting-pickup.css'
 })
+
 export class AwaitingPickup implements OnInit {
   awaitingPickup$?: Observable<any>
 
-  constructor(private dashboardService: UserDashboardService) {}
+  constructor(private dashboardService: Dashboard) {}
 
   ngOnInit(): void {
-    this.awaitingPickup$ = this.dashboardService.getUserAwaitingPickup()
- }
+    this.awaitingPickup$ = this.dashboardService.getAwaitingPickup()
+    AppConsole.log(this.awaitingPickup$)
+  }
 }
