--- frontend/src/app/user-dashboard/awaiting-pickup/awaiting-pickup.html
@@ -1,8 +1,0 @@
-<div class="w-full h-full flex items-center justify-center bg-gradient-to-br from-gray-100 to-gray-300 rounded-2xl shadow-xl backdrop-blur-lg border border-white/30 transition-all duration-300 hover:scale-105 hover:shadow-2xl hover:from-gray-200 hover:to-gray-400">
-    <div class="flex flex-col items-center justify-center w-full h-full p-8">
-        <span class="text-2xl font-semibold text-gray-800 mb-4 drop-shadow-lg transition-all duration-300 hover:text-gray-900 hover:scale-110">Unclaimed Parcel</span>
-        <span
-        *ngIf="awaitingPickup$ | async as ap"
-         class="text-6xl font-bold text-gray-900 bg-white/40 rounded-xl px-8 py-4 shadow-lg backdrop-blur-md border border-white/20 transition-all duration-300 hover:bg-white/60 hover:shadow-2xl hover:translate-y-[-4px]">{{ap.count}}</span>
-    </div>
-</div>
