--- frontend/src/app/user-dashboard/user-dashboard-module.ts
@@ -1,14 +1,0 @@
-import { NgModule } from '@angular/core';
-import { CommonModule } from '@angular/common';
-
-import { UserDashboardRoutingModule } from './user-dashboard-routing-module';
-
-
-@NgModule({
-  declarations: [],
-  imports: [
-    CommonModule,
-    UserDashboardRoutingModule
-  ]
-})
-export class UserDashboardModule { }
