--- frontend/src/app/user-dashboard/user-dashboard-parent/user-dashboard-parent.html
@@ -1,6 +1,0 @@
-<p class="mb-2">Welcome user!</p>
-
-<div class="grid grid-cols-3 gap-4">
-    <app-awaiting-pickup></app-awaiting-pickup>
-</div>
-
